Modify Grid Manager
Add spawn grid as a editor button. This is the current way to spawn the grid since spawning actors can't be done in constructor. May be possible later changed on spawning grid on variable change (GridTileCount).
Grid is spawned in editor so we can make additional changes on the tiles it spawns later. For example defining walls and tiles that are impossible to walk on and entrance where player enters the grid.

Also remove grid preview since spawning the grid is done the same way. The current size of the grid that are going to be used in the game is not going to exceed the amount that grid preview would be needed to preview the grid location.