Add character player functionality for entrance collision
Adds entrance collision for characters. When character goes inside of the collision character gets linked into the grid and movement is changed into the grid movement type. Also fix few things related to tile highlight.

A lot of the things set in entrance collision are placeholder. Check the file for further reasons for it.